using part directive and adding some simple docs
